Adding miros_code_path feature.
The client can specify a Path|str for the miros code.  If this is
specified the resulting miros code will be placed there.

If the client specifies a bool as the ``miros_code_path`` argument of
the ``XmlToMiro`` constructor, the default code path will be used (same
stem name and directory as the scxml path) but the file will not be
deleted once it has been used to specify a statechart.