Removed references to fns in all state transitions.
By calling the contents of the fns diction rather than the function
directly, the code was very difficult to debug, because I couldn't grep
and search for the functions that were causing a problem.  I can re-add
this stuff later, but for now we will keep the file as flat as possible
to make the research and debugging phase happen faster.