Changed structure of state functions.
The state functions now short-circuit return information about their
super state right away if they are called with a
SEARCH_FOR_SUPER_SIGNAL.  This is because any region can call them while
they are searching the graph.  An immediate return with the information
they are asking for is better than navigating through the full handler
if-else structure to get to the else clause.

I also noticed that when I'm looking at the code I am assuming the state
function is being called from its region only.  This isn't true, to
putting a return at the top from when it is not being called from its
region is a bit of defensive programming which will also speed up the
program.