Adjustment to tests to make the python version agnostic
The dill feature (serialized object) was breaking after a python
upgrade.  This was because the serialized object was made by hand
outside of the test flow.  The old version serialized under and older
version of python failed when I tried to deserialize it in a later
version of python.  To fix this I just have the test create the
serialized object every time it is run.